+++ spinta/backends/postgresql/commands/init.py
@@ -1,3 +1,4 @@
+from typing import overload
 
 import sqlalchemy as sa
 
@@ -7,7 +8,7 @@ from spinta import commands
 from spinta.components import Context, Model
 from spinta.core.enums import Access
 from spinta.manifests.components import Manifest
-from spinta.types.datatype import DataType, PrimaryKey
+from spinta.types.datatype import DataType, PrimaryKey, Ref
 from spinta.backends.constants import TableType
 from spinta.backends.helpers import get_table_name
 from spinta.backends.postgresql.constants import UNSUPPORTED_TYPES
@@ -15,8 +16,10 @@ from spinta.backends.postgresql.components import PostgreSQL
 from spinta.backends.postgresql.helpers import get_pg_name
 from spinta.backends.postgresql.helpers import get_column_name
 from spinta.backends.postgresql.helpers.changes import get_changes_table
+from spinta.datasets.enums import Level
 
 
+@overload
 @commands.prepare.register(Context, PostgreSQL, Manifest)
 def prepare(context: Context, backend: PostgreSQL, manifest: Manifest):
     # Prepare backend for models.
@@ -25,6 +28,7 @@ def prepare(context: Context, backend: PostgreSQL, manifest: Manifest):
             commands.prepare(context, backend, model)
 
 
+@overload
 @commands.prepare.register(Context, PostgreSQL, Model)
 def prepare(context: Context, backend: PostgreSQL, model: Model):
     columns = []
@@ -38,9 +42,20 @@ def prepare(context: Context, backend: PostgreSQL, model: Model):
             columns.extend(column)
         elif column is not None:
             columns.append(column)
+
     if model.unique:
-        for val in model.unique:
-            columns.append(sa.UniqueConstraint(*val))
+        for constraint in model.unique:
+            prop_list = []
+            for prop in constraint:
+                name = prop.name
+                if isinstance(prop.dtype, Ref):
+                    if prop.level is None or prop.level > Level.open:
+                        name = f'{name}._id'
+                    else:
+                        name = f'{name}.{prop.dtype.refprops[0].name}'
+                prop_list.append(name)
+            columns.append(sa.UniqueConstraint(*prop_list))
+
     # Create main table.
     main_table_name = get_pg_name(get_table_name(model))
     pkey_type = commands.get_primary_key_type(context, backend)
@@ -57,6 +72,8 @@ def prepare(context: Context, backend: PostgreSQL, model: Model):
     changelog_table = get_changes_table(context, backend, model)
     backend.add_table(changelog_table, model, TableType.CHANGELOG)
 
+
+@overload
 @commands.prepare.register(Context, PostgreSQL, DataType)
 def prepare(context: Context, backend: PostgreSQL, dtype: DataType):
     if dtype.name in UNSUPPORTED_TYPES:
@@ -69,6 +86,7 @@ def prepare(context: Context, backend: PostgreSQL, dtype: DataType):
         'date': sa.Date,
         'time': sa.Time,
         'datetime': sa.DateTime,
+        'temporal': sa.DateTime,
         'integer': sa.Integer,
         'number': sa.Float,
         'boolean': sa.Boolean,
@@ -95,6 +113,7 @@ def get_primary_key_type(context: Context, backend: PostgreSQL):
     return UUID()
 
 
+@overload
 @commands.prepare.register(Context, PostgreSQL, PrimaryKey)
 def prepare(context: Context, backend: PostgreSQL, dtype: PrimaryKey):
     pkey_type = commands.get_primary_key_type(context, backend)