+++ spinta/backends/postgresql/helpers/extractors.py
@@ -1,9 +1,10 @@
 import re
+from typing import List
 
 
-def extract_error_property_name(error_message: str) -> str:
+def extract_error_property_names(error_message: str) -> List[str]:
     pattern = re.compile(r'Key \((.*?)\)')
-    return pattern.search(error_message).group(1).split(".")[0]
+    return pattern.search(error_message).group(1).split(".")[0].split(', ')
 
 
 def extract_error_ref_id(error_message: str) -> str: