+++ tests/test_inspect.py
@@ -1,3 +1,4 @@
+import json
 import os
 import pathlib
 import tempfile
@@ -1594,4 +1595,170 @@ def test_inspect_with_manifest_backends(
          |   |   | name     | string  |      | NAME    |         | open    | Country name
          |   |   | code     | string  |      |         |         | open    | Country code
 ''')
+    assert a == b
+
+
+def test_inspect_json_model_ref_change(
+    rc: RawConfig,
+    cli: SpintaCliRunner,
+    tmp_path: Path):
+    json_manifest = [
+        {
+            "name": "Lithuania",
+            "code": "LT",
+            "location": {
+                "latitude": 54.5,
+                "longitude": 12.6
+            },
+            "cities": [
+                {
+                    "name": "Vilnius",
+                    "weather": {
+                        "temperature": 24.7,
+                        "wind_speed": 12.4
+                    }
+                },
+                {
+                    "name": "Kaunas",
+                    "weather": {
+                        "temperature": 29.7,
+                        "wind_speed": 11.4
+                    }
+                }
+            ]
+        },
+        {
+            "name": "Latvia",
+            "code": "LV",
+            "cities": [
+                {
+                    "name": "Riga"
+                }
+            ]
+        }
+    ]
+    path = tmp_path / 'manifest.json'
+    path.write_text(json.dumps(json_manifest))
+
+    result_file_path = tmp_path / 'result.csv'
+    # Configure Spinta.
+    rc = configure(rc, None, tmp_path / 'manifest.csv', f'''
+           d | r | m      | property            | type                   | ref    | source              
+           datasets/json/inspect                |                        |        |
+             | resource                         | json                   |        | {path}
+                                                |                        |        |
+             |   | Pos    |                     |                        | code   | .
+             |   |        | name                | string required unique |        | name
+             |   |        | code                | string required unique |        | code
+             |   |        | location_latitude   | number unique          |        | location.latitude
+             |   |        | location_longitude  | number unique          |        | location.longitude
+                                                |                        |        |
+             |   | Cities |                     |                        |        | cities
+             |   |        | name                | string required unique |        | name
+             |   |        | weather_temperature | number unique          |        | weather.temperature
+             |   |        | weather_wind_speed  | number unique          |        | weather.wind_speed
+           ''')
+
+    cli.invoke(rc, [
+        'inspect',
+        tmp_path / 'manifest.csv',
+        '-o', tmp_path / 'result.csv',
+    ])
+    # Check what was detected.
+    manifest = load_manifest(rc, result_file_path)
+    manifest.datasets['datasets/json/inspect'].resources['resource'].external = 'resource.json'
+    a, b = compare_manifest(manifest, f'''
+d | r | model  | property            | type                   | ref    | source
+datasets/json/inspect           |                        |        |
+  | resource                    | json                   |        | resource.json
+                                |                        |        |
+  |   | Pos                     |                        | code   | .
+  |   |   | name                | string required unique |        | name
+  |   |   | code                | string required        |        | code
+  |   |   | location_latitude   | number unique          |        | location.latitude
+  |   |   | location_longitude  | number unique          |        | location.longitude
+                                |                        |        |
+  |   | Cities                  |                        |        | cities
+  |   |   | name                | string required unique |        | name
+  |   |   | weather_temperature | number unique          |        | weather.temperature
+  |   |   | weather_wind_speed  | number unique          |        | weather.wind_speed
+  |   |   | parent              | ref                    | Pos    | ..
+    ''')
+    assert a == b
+
+
+def test_inspect_xml_model_ref_change(
+    rc: RawConfig,
+    cli: SpintaCliRunner,
+    tmp_path: Path):
+    xml = '''
+    <countries>
+        <country name="Lithuania" code="LT">
+            <location latitude="54.5" longitude="12.6"/>
+            <city name="Vilnius">
+                <weather>
+                    <temperature>24.7</temperature>
+                    <wind_speed>12.4</wind_speed>
+                </weather>
+            </city>
+            <city name="Kaunas">
+                <weather>
+                    <temperature>29.7</temperature>
+                    <wind_speed>11.4</wind_speed>
+                </weather>
+            </city>
+        </country>
+        <country name="Latvia" code="LV">
+            <city name="Riga"/>
+            <city name="Test"/>
+        </country>
+    </countries>
+'''
+    path = tmp_path / 'manifest.xml'
+    path.write_text(xml)
+
+    result_file_path = tmp_path / 'result.csv'
+    # Configure Spinta.
+    rc = configure(rc, None, tmp_path / 'manifest.csv', f'''
+           d | r | m      | property             | type                   | ref    | source              
+           datasets/xml/inspect                  |                        |        |
+             | resource                          | xml                    |        | {path}
+                                                 |                        |        |
+             |   | Country |                     |                        | code   | /countries/country
+             |   |         | name                | string required unique |        | @name
+             |   |         | code                | string required unique |        | @code
+             |   |         | location_latitude   | number unique          |        | location/@latitude
+             |   |         | location_longitude  | number unique          |        | location/@longitude
+                                                 |                        |        |
+             |   | City    |                     |                        |        | /countries/country/city
+             |   |         | name                | string required unique |        | @name
+             |   |         | weather_temperature | number unique          |        | weather/temperature
+             |   |         | weather_wind_speed  | number unique          |        | weather/wind_speed
+           ''')
+
+    cli.invoke(rc, [
+        'inspect',
+        tmp_path / 'manifest.csv',
+        '-o', tmp_path / 'result.csv',
+    ])
+    # Check what was detected.
+    manifest = load_manifest(rc, result_file_path)
+    manifest.datasets['datasets/xml/inspect'].resources['resource'].external = 'resource.xml'
+    a, b = compare_manifest(manifest, f'''
+d | r | model  | property            | type                   | ref    | source
+datasets/xml/inspect            |                        |        |
+  | resource                    | xml                    |        | resource.xml
+                                |                        |        |
+  |   | Country                 |                        | code   | /countries/country
+  |   |   | name                | string required unique |        | @name
+  |   |   | code                | string required        |        | @code
+  |   |   | location_latitude   | number unique          |        | location/@latitude
+  |   |   | location_longitude  | number unique          |        | location/@longitude
+                                |                        |        |
+  |   | City                    |                        |        | /countries/country/city
+  |   |   | name                | string required unique |        | @name
+  |   |   | weather_temperature | number unique          |        | weather/temperature
+  |   |   | weather_wind_speed  | number unique          |        | weather/wind_speed
+  |   |   | country             | ref                    | Country | ..
+    ''')
     assert a == b