+++ CHANGES.rst
@@ -3,7 +3,85 @@
 Changes
 #######
 
-0.1.50 (unreleased)
+0.1.54 (unreleased)
+===================
+
+
+0.1.53 (2023-08-01)
+===================
+
+New features:
+
+- Add support for RDF as manifest format (`#336`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/336
+
+- Add support for XML as manifest format (`#89`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/89
+
+Improvements:
+
+- Delete push target objects in correct order (`#458`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/458
+
+Bug fixes:
+
+- Add support for Oracle RAW type (`#493`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/493
+
+
+0.1.52 (2023-06-21)
+===================
+
+Improvements:
+
+- Recognize Oracle ROWID data type.
+
+
+0.1.51 (2023-06-20)
+===================
+
+New features:
+
+- Add support for `param` dimension (`#210`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/210
+
+- Spinta inspect now supports JSON data as schema source (`#98`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/98
+
+
+Improvements:
+
+- Recognize CHAR and BYTES data types (`#469`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/469
+
+
+- Allow writing data to models with base (`#205`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/205
+
+
+Bug fixes:
+
+- Fix spint push with ref type set to level 3 or below (`#460`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/460
+
+
+- Automatically add unique constraints for all primary keys specified in
+  model.ref (`#371`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/371
+
+
+
+0.1.50 (2023-05-22)
 ===================
 
 New features:
@@ -12,12 +90,40 @@ New features:
 
   __ https://github.com/atviriduomenys/spinta/issues/273
 
+- Add support for `unique` constraints in tabular manifests (`#148`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/148
+
+Improvements:
+
+- Much better implementation for updating manifest files from SQL as data
+  source (`#346`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/364
+
+- Show better error messages on foreign key constraint errors (`#363`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/363
+
+- Return a non-zero error code if `spinta push` command fails with an error
+  (`#423`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/423
+
+- Add support for older SQLite versions (`#411`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/411
+
 Bug fixes:
 
 - Correctly handle level 3 references, when referenced model does not have a
   primary key or property references a non-primary key (`#400`__).
 
   __ https://github.com/atviriduomenys/spinta/issues/400
+
+- WIPE command now works on tables with long names (`#431`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/431
 
 
 0.1.49 (2023-04-19)